+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 4.0.2
 - All: Add Geocoding of Address to Position Coordinates
 - Add Admin/SubAdmin area to Address
+- Add Distance Utilities into Abstractions
 
 
 ### 4.0.1